move types to shared lib
