+++ Composer/packages/client/src/pages/skills/skill-settings.tsx
@@ -6,8 +6,7 @@ import { jsx } from '@emotion/core';
 import React, { useState, useEffect } from 'react';
 import formatMessage from 'format-message';
 import { TextField } from 'office-ui-fabric-react/lib/TextField';
-
-import { DialogSetting } from '../../recoilModel/types';
+import { DialogSetting } from '@bfc/shared';
 
 import { FormFieldAlignHorizontalBotSettings } from './styles';
