+++ Composer/packages/client/src/recoilModel/persistence/FilePersistence.ts
@@ -3,13 +3,9 @@
 import keys from 'lodash/keys';
 import differenceWith from 'lodash/differenceWith';
 import isEqual from 'lodash/isEqual';
-import { DialogInfo } from '@bfc/shared';
+import { DialogInfo, DialogSetting, SkillManifest, BotAssets } from '@bfc/shared';
 
-import { DialogSetting } from '../../recoilModel/types';
-
-import { SkillManifest } from './../../pages/design/exportSkillModal/constants';
 import { LuFile, LgFile } from './../../../../lib/shared/src/types/indexers';
-import { BotAssets } from './../types';
 import * as client from './http';
 import { IFileChange, ChangeType, FileExtensions } from './types';
