+++ Composer/packages/client/src/recoilModel/persistence/__test__/FilePersistence.test.ts
@@ -1,8 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-import { DialogInfo, LgFile, LuFile } from '@bfc/shared';
+import { DialogInfo, LgFile, LuFile, BotAssets } from '@bfc/shared';
 
-import { BotAssets } from '../../types';
 import filePersistence from '../FilePersistence';
 
 jest.mock('axios', () => {
