+++ Composer/packages/lib/shared/src/types/indexers.ts
@@ -143,3 +143,58 @@ export interface SkillManifestInfo {
   lastModified: string;
   id: string;
 }
+
+export interface SkillManifest {
+  content: any;
+  id: string;
+  path?: string;
+  lastModified?: string;
+}
+
+export interface ILuisConfig {
+  name: string;
+  authoringKey: string;
+  endpointKey: string;
+  endpoint: string;
+  authoringEndpoint: string;
+  authoringRegion: string | 'westus';
+  defaultLanguage: string | 'en-us';
+  environment: string | 'composer';
+}
+
+export interface PublishTarget {
+  name: string;
+  type: string;
+  configuration: string;
+  lastPublished?: Date;
+}
+
+export interface DialogSetting {
+  MicrosoftAppId?: string;
+  MicrosoftAppPassword?: string;
+  luis: ILuisConfig;
+  publishTargets?: PublishTarget[];
+  runtime: {
+    customRuntime: boolean;
+    path: string;
+    command: string;
+  };
+  defaultLanguage: string;
+  languages: string[];
+  skill?: {
+    name: string;
+    manifestUrl: string;
+  }[];
+  botId?: string;
+  skillHostEndpoint?: string;
+  [key: string]: unknown;
+}
+
+export type BotAssets = {
+  projectId: string;
+  dialogs: DialogInfo[];
+  luFiles: LuFile[];
+  lgFiles: LgFile[];
+  skillManifests: SkillManifest[];
+  setting: DialogSetting;
+};
